+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This is the first release. It continues from ChaosTools.jl v2.9, and hence, the 
 - Added a new clause in automatic `ε` estimation in `AttractorsViaProximity` for when there is only a single attractor passed in by the user.
 
 ## Basin fractions continuation
-- New function `basins_fractions_continuation` that calculates basins fractions and how these change versus a parameter (given a continuation method)
+- New function `continuation` that calculates basins fractions and how these change versus a parameter (given a continuation method)
 - New basins fraction continuation method `RecurrencesSeededContinuation` that utilizes a brand new algorithm to continuate basins fractions of arbitrary systems.
 - `match_attractor_ids!` has been fully overhauled to be more flexible, allow more ways to match, and also allow arbitrary user-defined ways to match.
 - New function `match_basins_ids!` for matching the output of basins_of_attraction`.